Mod Update 2.74
Версия 2.74: вырезан кастомный шрифт, смайлы взял из ВК (можно отключить системные смайлы). Голосовухи можно расшифровать. Фикс двухэтапной авторизации

+++ app/src/main/java/biz/dealnote/messenger/api/model/VKApiAudioPlaylist.java
@@ -29,7 +29,7 @@ public class VKApiAudioPlaylist {
     public int id;
     public int owner_id;
     public int count;
-    public int update_time;
+    public long update_time;
     public int Year;
     public String artist_name;
     public String genre;
